+++ src/Promenade.Service/EmediaService.cs
@@ -19,7 +19,10 @@ namespace Ocuda.Promenade.Service
         private readonly ICategoryTextRepository _categoryTextRepository;
         private readonly IEmediaRepository _emediaRepository;
         private readonly IEmediaCategoryRepository _emediaCategoryRepository;
+        private readonly IEmediaGroupRepository _emediaGroupRepository;
         private readonly IEmediaTextRepository _emediaTextRepository;
+        private readonly ISegmentRepository _segmentRepository;
+        private readonly ISegmentTextRepository _segmentTextRepository;
         private readonly LanguageService _languageService;
 
         public EmediaService(ILogger<EmediaService> logger,
@@ -29,7 +32,10 @@ namespace Ocuda.Promenade.Service
             ICategoryTextRepository categoryTextRepository,
             IEmediaRepository emediaRepository,
             IEmediaCategoryRepository emediaCategoryRepository,
+            IEmediaGroupRepository emediaGroupRepository,
             IEmediaTextRepository emediaTextRepository,
+            ISegmentRepository segmentRepository,
+            ISegmentTextRepository segmentTextRepository,
             LanguageService languageService)
             : base(logger, dateTimeProvider)
         {
@@ -43,13 +49,19 @@ namespace Ocuda.Promenade.Service
                 ?? throw new ArgumentNullException(nameof(emediaRepository));
             _emediaCategoryRepository = emediaCategoryRepository
                 ?? throw new ArgumentNullException(nameof(emediaCategoryRepository));
+            _emediaGroupRepository = emediaGroupRepository
+                ?? throw new ArgumentNullException(nameof(emediaGroupRepository));
             _emediaTextRepository = emediaTextRepository
                 ?? throw new ArgumentNullException(nameof(emediaTextRepository));
+            _segmentRepository = segmentRepository
+                ?? throw new ArgumentNullException(nameof(segmentRepository));
+            _segmentTextRepository = segmentTextRepository
+                ?? throw new ArgumentNullException(nameof(segmentTextRepository));
             _languageService = languageService
                 ?? throw new ArgumentNullException(nameof(languageService));
         }
 
-        public async Task<List<Emedia>> GetAllEmediaAsync()
+        public async Task<ICollection<EmediaGroup>> GetGroupedEmediaAsync()
         {
             var currentCultureName = _httpContextAccessor
                 .HttpContext
@@ -59,61 +71,72 @@ namespace Ocuda.Promenade.Service
                 .UICulture?
                 .Name;
 
+            int defaultLanguageId = await _languageService.GetDefaultLanguageIdAsync();
             int? currentLangaugeId = null;
-            int? defaultLanguageId = null;
 
-            var emedias = await _emediaRepository.GetAllAsync();
-
-            foreach (var emedia in emedias)
+            if (!string.IsNullOrWhiteSpace(currentCultureName))
             {
-                if (!string.IsNullOrWhiteSpace(currentCultureName))
+                currentLangaugeId = await _languageService.GetLanguageIdAsync(currentCultureName);
+            }
+
+            var groupedEmedia = await _emediaGroupRepository.GetAllWithEmediaAsync();
+
+            foreach (var group in groupedEmedia)
+            {
+                group.Emedias = group.Emedias.OrderBy(_ => _.Name).ToList();
+
+                if (group.SegmentId.HasValue)
                 {
-                    if (!currentLangaugeId.HasValue)
+                    group.Segment = await _segmentRepository.GetActiveAsync(group.SegmentId.Value);
+
+                    if (group.Segment != null)
                     {
-                        currentLangaugeId = await _languageService
-                            .GetLanguageIdAsync(currentCultureName);
+                        if (currentLangaugeId.HasValue)
+                        {
+                            group.Segment.SegmentText = await _segmentTextRepository
+                                .GetByIdsAsync(currentLangaugeId.Value, group.Segment.Id);
+                        }
+                        if (group.Segment.SegmentText == null)
+                        {
+                            group.Segment.SegmentText = await _segmentTextRepository
+                                .GetByIdsAsync(defaultLanguageId, group.Segment.Id);
+                        }
                     }
-                    emedia.EmediaText = await _emediaTextRepository.GetByIdsAsync(
-                        emedia.Id, currentLangaugeId.Value);
-                }
-                if (emedia.EmediaText == null)
-                {
-                    if (!defaultLanguageId.HasValue)
-                    {
-                        defaultLanguageId = await _languageService.GetDefaultLanguageIdAsync();
-                    }
-                    emedia.EmediaText = await _emediaTextRepository.GetByIdsAsync(
-                        emedia.Id, defaultLanguageId.Value);
                 }
 
-                emedia.Categories = await _emediaCategoryRepository
-                    .GetCategoriesByEmediaIdAsync(emedia.Id);
-
-                foreach (var category in emedia.Categories)
+                foreach (var emedia in group.Emedias)
                 {
-                    if (!string.IsNullOrWhiteSpace(currentCultureName))
+                    if (currentLangaugeId.HasValue)
                     {
-                        if (!currentLangaugeId.HasValue)
-                        {
-                            currentLangaugeId = await _languageService
-                                .GetLanguageIdAsync(currentCultureName);
-                        }
-                        category.CategoryText = await _categoryTextRepository.GetByIdsAsync(
-                            category.Id, currentLangaugeId.Value);
+                        emedia.EmediaText = await _emediaTextRepository.GetByIdsAsync(
+                            emedia.Id, currentLangaugeId.Value);
                     }
-                    if (category.CategoryText == null)
+                    if (emedia.EmediaText == null)
                     {
-                        if (!defaultLanguageId.HasValue)
+                        emedia.EmediaText = await _emediaTextRepository.GetByIdsAsync(
+                            emedia.Id, defaultLanguageId);
+                    }
+
+                    emedia.Categories = await _emediaCategoryRepository
+                        .GetCategoriesByEmediaIdAsync(emedia.Id);
+
+                    foreach (var category in emedia.Categories)
+                    {
+                        if (currentLangaugeId.HasValue)
                         {
-                            defaultLanguageId = await _languageService.GetDefaultLanguageIdAsync();
+                            category.CategoryText = await _categoryTextRepository.GetByIdsAsync(
+                                category.Id, currentLangaugeId.Value);
                         }
-                        category.CategoryText = await _categoryTextRepository.GetByIdsAsync(
-                            category.Id, defaultLanguageId.Value);
+                        if (category.CategoryText == null)
+                        {
+                            category.CategoryText = await _categoryTextRepository.GetByIdsAsync(
+                                category.Id, defaultLanguageId);
+                        }
                     }
                 }
             }
 
-            return emedias;
+            return groupedEmedia;
         }
     }
 }